Move ItemStatus to the module Common;
Reword JSDoc;

Match the namespaced import style in the search test.
